Add DynamicList as a type
