+++ ccd-config-generator/src/test/java/uk/gov/hmcts/reform/fpl/model/HearingDateDynamicElement.java
@@ -3,9 +3,9 @@ package uk.gov.hmcts.reform.fpl.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
 import lombok.Data;
-import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicElementIndicator;
-import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicList;
-import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicListElement;
+import uk.gov.hmcts.ccd.sdk.type.DynamicElementIndicator;
+import uk.gov.hmcts.ccd.sdk.type.DynamicList;
+import uk.gov.hmcts.ccd.sdk.type.DynamicListElement;
 
 import java.util.UUID;
