+++ ccd-config-generator/src/test/java/uk/gov/hmcts/reform/fpl/utils/ElementUtils.java
@@ -1,8 +1,8 @@
 package uk.gov.hmcts.reform.fpl.utils;
 
+import uk.gov.hmcts.ccd.sdk.type.DynamicList;
+import uk.gov.hmcts.ccd.sdk.type.DynamicListElement;
 import uk.gov.hmcts.reform.fpl.model.common.Element;
-import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicList;
-import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicListElement;
 
 import java.util.Collection;
 import java.util.Collections;
